Populate the flights field of the booking details DTO
